Fix: Correct Apk Signing in GitHub Actions
This commit updates the Apk signing process in the GitHub Actions workflow. Previously, `sed` was used to modify `gradle.properties`, but it was not creating the file if it didn't exist. This change replaces the `sed` command with `echo` to ensure `gradle.properties` is created and the signing configuration is correctly written.

Specific changes:
- In `.github/workflows/build.yml`:
    - Changed the `Apk Sign` step to use `echo "signing_release_storeFileFromRoot=./bilibilias.jks" > $GITHUB_WORKSPACE/gradle.properties` instead of `sed`.
- In `.github/workflows/src.main.kts`:
    - Updated the `Apk Sign` step command similarly to use `echo` for creating/updating `gradle.properties`.


Former-commit-id: 7dc1d47c97b87d4e726dc3b8696f0e7455f9d29c